requirements.txt
python-dotenv==0.21.0
python-telegram-bot==13.14
googletrans==4.0.0-rc1